% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/showtext.R
\name{showtext.end}
\alias{showtext.end}
\title{Turn Off 'showtext' Text Rendering}
\usage{
showtext.end()
}
\description{
This function will turn off the \pkg{showtext} functionality
of rendering text. When you call this function, the current
active device should be the same with the one when you call
\code{\link{showtext.begin}()}, or an error will be issued.
See the example in \code{\link{showtext.begin}()}.
}
\author{
Yixuan Qiu <\url{http://yixuan.cos.name/}>
}
\seealso{
\code{\link{showtext.begin}()}
}

