% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showtext.R
\name{showtext_end}
\alias{showtext_end}
\alias{showtext.end}
\title{Turning Off 'showtext' Text Rendering}
\usage{
showtext_end()

showtext.end()
}
\description{
The two versions of this function are equivalent, but the
"underscore" naming is preferred.

This function will turn off the \pkg{showtext} functionality
of rendering text. When you call this function, the current
active device should be the same as the one when you call
\code{\link{showtext_begin}()}, or an error will be issued.
See the example in \code{\link{showtext_begin}()}.
}
\seealso{
\code{\link{showtext_begin}()}
}
\author{
Yixuan Qiu <\url{https://statr.me/}>
}
