\name{showtext.end}
\alias{showtext.end}
\title{Turn off "showtext" text rendering}
\usage{
showtext.end()
}
\description{
This function will turn off the \pkg{showtext}
functionality of rendering text. When you call this
function, the current active device should be the same with
the one when you call \code{\link{showtext.begin}()}, or an
error will be issued. See the example in
\code{\link{showtext.begin}()}.
}
\author{
Yixuan Qiu <\url{http://yixuan.cos.name/}>
}
\seealso{
\code{\link{showtext.begin}()}
}

